New hydration/dehydration procedure
Dehydration now collects all instance variables, checks whether the SQL table contains columns named accordingly and then dehydrates the present values. What is now needed is a fast dehydration, e.g. to be used in a loop, which doesn't check the table structure on each call and which allows partial dehydrations so that non-hydrated values don't get set to null.